basic plugin mechanism, .i18nrc file, fixes #12
this should eliminate the need for most (all?) glue code... so now in your
project, you can create an .i18nrc file containing json like so:

    {
      "inferredKeyFormat": "underscored",
      "plugins": [
        "react-i18nliner"
      ]
    }

this allows you to override settings (e.g. inferredKeyFormat) and specify
plugins. for each plugin, i18nliner will require it (it expects a
function), and will call it with a single arg, an object that allows the
plugin to register itself. currently the only property of this object is
a "processors" object, which the plugin can mutate to extend or add
processors (i18nliner will run all processors in this object).

for example, i18nliner-handlebars should be reworked to export a single
function along these lines:

    module.exports = function(i18nliner) {
      i18nliner.processors.HbsProcessor = HbsProcessor;
    }